аааааааааа

+++ source/psychlua/TextFunctions.hx
@@ -2,11 +2,13 @@ package psychlua;
 
 class TextFunctions
 {
+	static var lua:State;
+
 	public static function implement(funk:FunkinLua)
 	{
-		var lua = funk.lua;
+		lua = funk.lua;
 		var game:PlayState = PlayState.instance;
-		Lua_helper.add_callback(lua, "makeLuaText", function(tag:String, text:String, width:Int, x:Float, y:Float) {
+		addCallback("makeLuaText", function(tag:String, text:String, width:Int, x:Float, y:Float) {
 			tag = tag.replace('.', '');
 			LuaUtils.resetTextTag(tag);
 			var leText:FlxText = new FlxText(x, y, width, text, 16);
@@ -17,7 +19,7 @@ class TextFunctions
 			game.modchartTexts.set(tag, leText);
 		});
 
-		Lua_helper.add_callback(lua, "setTextString", function(tag:String, text:String) {
+		addCallback("setTextString", function(tag:String, text:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -27,7 +29,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextString: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextSize", function(tag:String, size:Int) {
+		addCallback("setTextSize", function(tag:String, size:Int) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -37,7 +39,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextSize: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextWidth", function(tag:String, width:Float) {
+		addCallback("setTextWidth", function(tag:String, width:Float) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -47,7 +49,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextWidth: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextBorder", function(tag:String, size:Int, color:String) {
+		addCallback("setTextBorder", function(tag:String, size:Int, color:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -64,7 +66,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextBorder: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextColor", function(tag:String, color:String) {
+		addCallback("setTextColor", function(tag:String, color:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -74,7 +76,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextColor: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextFont", function(tag:String, newFont:String) {
+		addCallback("setTextFont", function(tag:String, newFont:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -84,7 +86,7 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextFont: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextItalic", function(tag:String, italic:Bool) {
+		addCallback("setTextItalic", function(tag:String, italic:Bool) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
@@ -94,17 +96,15 @@ class TextFunctions
 			FunkinLua.luaTrace("setTextItalic: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return false;
 		});
-		Lua_helper.add_callback(lua, "setTextAlignment", function(tag:String, alignment:String = 'left') {
+		addCallback("setTextAlignment", function(tag:String, alignment:String = 'left') {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
 			if(obj != null)
 			{
-				obj.alignment = LEFT;
 				switch(alignment.trim().toLowerCase())
 				{
-					case 'right':
-						obj.alignment = RIGHT;
-					case 'center':
-						obj.alignment = CENTER;
+					case 'right':	obj.alignment = RIGHT;
+					case 'center':	obj.alignment = CENTER;
+					default:		obj.alignment = LEFT;
 				}
 				return true;
 			}
@@ -112,58 +112,42 @@ class TextFunctions
 			return false;
 		});
 
-		Lua_helper.add_callback(lua, "getTextString", function(tag:String) {
+		addCallback("getTextString", function(tag:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
-			if(obj != null && obj.text != null)
-			{
-				return obj.text;
-			}
+			if(obj != null && obj.text != null) return obj.text;
 			FunkinLua.luaTrace("getTextString: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return null;
 		});
-		Lua_helper.add_callback(lua, "getTextSize", function(tag:String) {
+		addCallback("getTextSize", function(tag:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
-			if(obj != null)
-			{
-				return obj.size;
-			}
+			if(obj != null) return obj.size;
 			FunkinLua.luaTrace("getTextSize: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return -1;
 		});
-		Lua_helper.add_callback(lua, "getTextFont", function(tag:String) {
+		addCallback("getTextFont", function(tag:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
-			if(obj != null)
-			{
-				return obj.font;
-			}
+			if(obj != null) return obj.font;
 			FunkinLua.luaTrace("getTextFont: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return null;
 		});
-		Lua_helper.add_callback(lua, "getTextWidth", function(tag:String) {
+		addCallback("getTextWidth", function(tag:String) {
 			var obj:FlxText = LuaUtils.getTextObject(tag);
-			if(obj != null)
-			{
-				return obj.fieldWidth;
-			}
+			if(obj != null) return obj.fieldWidth;
 			FunkinLua.luaTrace("getTextWidth: Object " + tag + " doesn't exist!", false, false, FlxColor.RED);
 			return 0;
 		});
 
-		Lua_helper.add_callback(lua, "addLuaText", function(tag:String) {
+		addCallback("addLuaText", function(tag:String) {
 			if(game.modchartTexts.exists(tag)) {
 				var shit:FlxText = game.modchartTexts.get(tag);
 				LuaUtils.getTargetInstance().add(shit);
 			}
 		});
-		Lua_helper.add_callback(lua, "removeLuaText", function(tag:String, destroy:Bool = true) {
-			if(!game.modchartTexts.exists(tag)) {
-				return;
-			}
+		addCallback("removeLuaText", function(tag:String, destroy:Bool = true) {
+			if(!game.modchartTexts.exists(tag)) return;
 
 			var pee:FlxText = game.modchartTexts.get(tag);
-			if(destroy) {
-				pee.kill();
-			}
+			if(destroy) pee.kill();
 
 			LuaUtils.getTargetInstance().remove(pee, true);
 			if(destroy) {
@@ -172,4 +156,6 @@ class TextFunctions
 			}
 		});
 	}
+
+	static function addCallback(name:String, func:Dynamic) Lua_helper.add_callback(lua, name, func);
 }
